+++ app/src/main/java/com/example/aibudgetapp/ui/screens/budget/TransactionsScreen.kt
@@ -1,0 +1,23 @@
+package com.example.aibudgetapp.ui.screens.budget
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TransactionsScreen() {
+    Column {
+        Text("🧾 Recent Transactions", style = MaterialTheme.typography.titleMedium)
+        Spacer(modifier = Modifier.height(8.dp))
+        Text("- Starbucks Coffee: $5.50")
+        Text("- Netflix Subscription: $12.99")
+        Text("- Gas Station: $40.00")
+    }
+}
